Update google_webfonts.py
Added working functionality for merge_fonts command. Works on link tags added by add_fonts command. Although it works, code is far from pythonic (and not at all modular). Merge_fonts doesn't yet check for multiple weights from the same font-family
e.g. if there is a link tag for roboto:200 and a link tag for roboto:500, this will merge to roboto:200|roboto:500, this should  merge to roboto:200,500. 
A possible fix for this: analyze the filtered link parts with regex and scan for doubles, then organize them by font-family in a dictionary and build a new link tag from scratch.